docs(readme): remove contribution sections
Unlike other repositories in the component system, the interfaces
are at the core of the data model of the system. For this reason,
we should not accept external contributions, but only core members
should make the appropriate changes after a design has been agreed on.
This could be with either, new features developed by core members,
or design in conjunction with external contributors.