build(rollup): add ESM bundle configurations
This also simplifies (and makes a slightly smaller) bundle by removing Babel transpilation.
It is assumed that if an application is consuming the library as an ES module, then it will most likely take care of transpilation when bundling everything together.
Therefore, there's no need for the interface library to add Babel code